Also validate range for `NumberInlineInput`
